+++ CHANGES.markdown
@@ -24,6 +24,7 @@ Bug fixes and minor enhancements.
 * Mash editor needs some tidy-up [963](https://github.com/Brewtarget/brewtarget/issues/963)
 * Better Hop Stage default for hop additions [1018](https://github.com/Brewtarget/brewtarget/issues/1018)
 * Qt Libs missing in 4.2.0 on MacOS [1022](https://github.com/Brewtarget/brewtarget/issues/1022)
+* XML recipe export doesn't work when no file extension supplied [1025](https://github.com/Brewtarget/brewtarget/issues/1025)
 
 ### Release Timestamp
 Sun, 24 Aug 2025 04:02:01 +0100